Clean up implicit conversions (#61)
* Clean up implicit conversions

Get rid of a bunch of warnings

* Get rid of choc and comment out unused variables in cnpy

+++ NeuralAmpModeler/dsp/Resample.h
@@ -55,7 +55,7 @@ void dsp::ResampleCubic(const std::vector<T> &inputs,
   while (time < endTimeOriginal) {
     // Find the index of the sample in the original audio file that is just
     // before the current time in the resampled audio file
-    int index = (long)std::floor(time / timeIncrement);
+    long index = (long)std::floor(time / timeIncrement);
 
     // Calculate the time difference between the current time in the resampled
     // audio file and the sample in the original audio file
